+++ media/media-card/src/utils/lightCards/cardLoading.tsx
@@ -1,24 +1,30 @@
 import React from 'react';
-import { Component } from 'react';
+import SpinnerIcon from '@atlaskit/spinner';
+import { expValEquals } from '@atlaskit/tmp-editor-statsig/exp-val-equals';
+import { LoadingBar } from '../../card/ui/loadingBar/loadingBar';
 import { getDimensionsWithDefault } from './getDimensionsWithDefault';
 import { Wrapper } from './lightCardWrappers';
 import { type StaticCardProps } from './types';
-import SpinnerIcon from '@atlaskit/spinner';
 
-export class CardLoading extends Component<StaticCardProps, {}> {
-	render(): React.JSX.Element {
-		const { dimensions: dimensionsProp, testId, interactionName } = this.props;
-		const dimensions = getDimensionsWithDefault(dimensionsProp);
+export const CardLoading = ({
+	dimensions: dimensionsProp,
+	testId,
+	interactionName,
+}: StaticCardProps): React.JSX.Element => {
+	const dimensions = getDimensionsWithDefault(dimensionsProp);
 
-		return (
-			<Wrapper
-				data-testid={testId || 'media-card-loading'}
-				data-test-loading
-				dimensions={dimensions}
-				data-vc="media-card-loading"
-			>
+	return (
+		<Wrapper
+			data-testid={testId || 'media-card-loading'}
+			data-test-loading
+			dimensions={dimensions}
+			data-vc="media-card-loading"
+		>
+			{expValEquals('cc-maui-ai-edit-loading-experiment', 'isEnabled', true) ? (
+				<LoadingBar interactionName={interactionName || 'media-card-loading'} />
+			) : (
 				<SpinnerIcon interactionName={interactionName || 'media-card-loading'} />
-			</Wrapper>
-		);
-	}
-}
+			)}
+		</Wrapper>
+	);
+};